Allow newer aws provider versions (#86)
Thank you for contribution, @saulius ! And sorry for the delay.
Tests are failing for some reason, but it looks like it’s unrelated to your change. I will figure it out before releasing a new tag. In the meantime, you can use the latest version from the `master` branch. This is not recommended in general, I hope I’ll figure out the problem with tests soon. I probably need to do some updates there as well e.g. use the latest version of Terraform for tests, etc.
Cc: @wielkiborsuk